+++ CHANGELOG.rst
@@ -42,6 +42,7 @@ Bug fixes and small improvements:
 - Fix text report for covered decisions. (:issue:`1192`)
 - Fix runtime problem introduced with 8.4. (:issue:`1194`)
 - Fix wrong entries in data source attribute of JSON report. (:issue:`1194`)
+- Fix nested HTML report without coverage data. (:issue:`1197`)
 
 Documentation:
 
@@ -82,6 +83,7 @@ Known bugs:
 - :ref:`fix_1192`
 - :ref:`fix_1194_1`
 - :ref:`fix_1194_2`
+- :ref:`fix_1197`
 
 Breaking changes:
 
@@ -194,6 +196,7 @@ Known bugs:
 - :ref:`fix_1138`
 - :ref:`fix_1171_1`
 - :ref:`fix_1176`
+- :ref:`fix_1197`
 
 Breaking changes:
 
@@ -257,6 +260,7 @@ Known bugs:
 - :ref:`fix_1130`
 - :ref:`fix_1138`
 - :ref:`fix_1171_1`
+- :ref:`fix_1197`
 
 Breaking changes:
 
@@ -285,6 +289,7 @@ Known bugs:
 - :ref:`fix_1130`
 - :ref:`fix_1138`
 - :ref:`fix_1171_1`
+- :ref:`fix_1197`
 
 Breaking changes:
 
@@ -317,6 +322,7 @@ Known bugs:
 - :ref:`fix_1130`
 - :ref:`fix_1138`
 - :ref:`fix_1171_1`
+- :ref:`fix_1197`
 
 Breaking changes:
 
@@ -397,6 +403,7 @@ Known bugs:
 - :ref:`fix_1037`
 - :ref:`fix_1089`.
 - :ref:`fix_1138`
+- :ref:`fix_1197`
 
 New features and notable changes:
 
@@ -431,6 +438,7 @@ Known bugs:
 - :ref:`fix_1022`
 - :ref:`fix_1037`
 - :ref:`fix_1089`
+- :ref:`fix_1197`
 
 Breaking changes:
 
@@ -532,6 +540,7 @@ Internal changes:
 Known bugs:
 
 - :ref:`fix_1037`
+- :ref:`fix_1197`
 
 Breaking changes: